refactor: simplify daily to not need a specific flag packages
instead, let the packages be the arguments passed.
I took the liberty of also not including a default,
however could make a "default" by just checking
if no args and then setting that value.

This also removes a global variable and makes
daily() easier to test as can pass it custom args